test(#1582): fold the five modules that were tested from two places
Five modules had a test file in BOTH of the locations being emptied
and in `src/__tests__/`, each collision resolved by hand here. The
issue named three; `channelTopic` and `mentionScroll` are the two it
missed, because it compared the two `__tests__/` directories with each
other and never compared either against the co-located files.

The rule applied, and it is deliberately narrow: a case disappears ONLY
if it is byte-identical to one that stays, or strictly weaker than one
that stays — and in that second case the STRICTER survives. "Reads like
a paraphrase of its neighbour" is not on that list, and measuring the
pairs is what showed why it must not be. `passkeys` looked like the
worst duplication of the five (two pairs the issue calls "the same
claim written twice") and turned out to have none: each member asserted
fields the other did not — `challenge` bytes on one side, `user.name`,
`attestation` and `excludeCredentials` on the other — so neither
dominated and both survive. Same for `mentionScroll`'s two below-the-
fold pairs, which differ only in fixture data.

So exactly TWO cases die, both on `customTheme.tokenToCssVars`: the
`mono-default` omission was asserted twice with byte-identical bodies,
and the named-family mapping twice differing only in strictness — the
copy asserting `toContain('"jetbrains-mono"')` is kept over the one
asserting it unquoted, because a regression that dropped the CSS
quoting would slip straight past the looser assertion.

6208 cases before, 6206 after. The evidence is not the arithmetic: it
is that diffing the full case-NAME set of the whole suite before
against after yields exactly two lines, both deletions, both the two
above — no case renamed, none invented. That is why no surviving case
was renamed even where a tidier name was available: keeping the name
diff at two lines is what makes the conservation auditable by anyone,
and it is worth more than two neater describe headings.

`channelTopic` and `mobilePanel` are the ones that mattered most. Their
APIs were PARTITIONED across the pair — store here, pure helpers there;
four launcher verbs here, three there — so a verb missing from the file
you opened told you nothing at all. Absence is evidence again.

Three merges also had to collapse incompatible plumbing, since one
module path admits one mock factory and one `beforeEach` chain:
  * mobilePanel — two different fakes of `../lib/archive` (a factory
    `vi.fn()` read through a dynamic import vs a closure-captured spy
    behind a wrapper) become one both halves read directly;
  * passkeys — `vi.stubGlobal("navigator", {})` gives way to defining
    and DELETING `navigator.credentials` on the real one, which is
    strictly more faithful: jsdom's baseline is an absent property and
    that is exactly what #725's plain-http detection probes for;
  * customTheme — two `payload()` builders seeding different colors
    become one, on the values the merged-in cases assert literally.
Each of these fails LOUDLY if it stops working (a non-spy handed to
`toHaveBeenCalledWith`, a ceremony that throws) rather than passing
vacuously, which is what makes green meaningful evidence here.

One hazard the merge created and had to close: `customTheme`'s day/night
boot cases REPLACE `window.matchMedia` by plain assignment, which
`vi.restoreAllMocks` cannot undo. In their own file that leaked no
further than the file; sharing a file with the #837 mount cases, which
read the real one, makes restoring it load-bearing rather than tidy.
