+++ cicchetto/src/__tests__/mentionScroll.test.ts
@@ -5,9 +5,19 @@ import {
   type ScrollbackLineGeom,
 } from "../lib/mentionScroll";
 
-// Pure geometry core for the #360 mention-aware scroll-to-bottom badge +
-// the mention JUMP anchor. No DOM: jsdom reports 0 for every layout read,
-// so the decision logic is exercised here against synthetic geometry.
+// #360 — pure geometry core for the mention-aware scroll-to-bottom badge plus
+// the mention JUMP anchor. jsdom is blind to real layout (offsetTop /
+// clientHeight read 0), so the DOM measuring lives in ScrollbackPane and the
+// DECISION is isolated here as pure functions over pre-measured geometry. The
+// e2e pins the DOM→scroll wiring in a real browser; these pin the predicate.
+//
+// #1582 — merged from a second file that tested this same module from
+// `src/lib/mentionScroll.test.ts`. The two overlapped on the below-the-fold
+// predicate without either dominating the other (different fixtures, same
+// claim), so BOTH survive: a case is dropped only when it is byte-identical to
+// one that stays or strictly weaker than one that stays. The near-duplicates
+// are placed ADJACENT rather than scattered, so the next reader can see the
+// redundancy and judge it instead of rediscovering it.
 
 const line = (id: number, top: number, isMention = false): ScrollbackLineGeom => ({
   id,
@@ -27,9 +37,52 @@ describe("mentionsBelowViewport", () => {
     expect(mentionsBelowViewport(lines, 300)).toEqual([3, 4]);
   });
 
+  it("returns nearest-first ids of mention lines entirely below the fold", () => {
+    const lines = [
+      line(1, 0, true), // above the fold — already seen
+      line(2, 100, false),
+      line(3, 300, true), // below the fold
+      line(4, 500, false),
+      line(5, 700, true), // below the fold
+    ];
+    // viewport bottom at 200px: lines with top >= 200 are below the fold.
+    expect(mentionsBelowViewport(lines, 200)).toEqual([3, 5]);
+  });
+
   it("excludes a mention straddling the fold (partially visible = seen)", () => {
     const lines = [line(1, 290, true)];
     // top 290 < viewportBottom 300 → straddling → excluded.
+    expect(mentionsBelowViewport(lines, 300)).toEqual([]);
+  });
+
+  it("excludes partially-visible mentions whose top is above the fold", () => {
+    // A mention straddling the fold (top above viewportBottom) is treated as
+    // seen — not counted, not a jump target.
+    const lines = [line(1, 150, true)];
+    expect(mentionsBelowViewport(lines, 200)).toEqual([]);
+  });
+
+  it("preserves DOM order so element[0] is the nearest jump target", () => {
+    const lines = [line(9, 400, true), line(7, 800, true), line(8, 1200, true)];
+    // Input order == chronological order; nearest-below is the smallest top,
+    // which is the first element. The consumer jumps to below[0].
+    expect(mentionsBelowViewport(lines, 100)).toEqual([9, 7, 8]);
+  });
+
+  it("excludes non-mention lines below the fold", () => {
+    const lines = [line(1, 300, false), line(2, 500, false)];
+    expect(mentionsBelowViewport(lines, 100)).toEqual([]);
+  });
+
+  it("counts a mention whose top sits exactly at the viewport bottom", () => {
+    // top === viewportBottom ⇒ the line begins exactly at the fold, so it is
+    // entirely below and unseen.
+    const lines = [line(1, 200, true)];
+    expect(mentionsBelowViewport(lines, 200)).toEqual([1]);
+  });
+
+  it("returns empty when every mention is above the fold", () => {
+    const lines = [line(1, 0, true), line(2, 50, true), line(3, 120, false)];
     expect(mentionsBelowViewport(lines, 300)).toEqual([]);
   });
 });
